Reformatting of autoria bs4 parser
Parser is moved to separate module - bs4_parser
Minor changes and improvements of parser and related code
