+++ bs4_parser/main.py
@@ -1,0 +1,9 @@
+from common import parse_autoria
+
+
+def main():
+    parse_autoria()
+
+
+if __name__ == '__main__':
+    main()
